refactor: Remove unnecessary codegen for UDFs and UDAFs. This was adding a lot of extra complexity.
