+++ ksql-engine/src/main/java/io/confluent/ksql/function/UdfInvoker.java
@@ -16,14 +16,12 @@
 package io.confluent.ksql.function;
 
 /**
- * This interface is used by the UDFCompiler to generate and compile
- * a class for each function that is loaded via the UDFLoader
+ * This interface is used to invoke UDFs
  */
 public interface UdfInvoker {
 
   /**
-   * Call onto an UDF instance with the expected args. An implementation of this
-   * is generated by the UDFCompiler with the appropriate args. So this is providing
+   * Call onto an UDF instance with the expected args. This is providing
    * a wrapper such that we can invoke all UDFs in a generic way.
    *
    * @param udf     the udf that is being called
